TwiWatchdogCount jetzt 16 bit.
git-svn-id: svn://svn.code.sf.net/p/itelex/misc-code/trunk@99 08b3b9a4-00d7-49ad-bbbd-872362c6a4d0

+++ Gemeinsam/BusKomm.c
@@ -37,7 +37,7 @@ volatile uint8_t BusKollisionZaehler;
 volatile bool BusEmpfMark;
 volatile bool BusEmpfMarkwechsel;
 volatile uint16_t TwiIsrCount;
-volatile uint8_t TwiWatchdogCount;
+volatile uint16_t TwiWatchdogCount;
 
 // lokal:
 static volatile uint8_t BusEmpfPuffer[EMPF_PUFFER_GROESSE];
